Fix crash when the number of columns is 1 and a 0 width character is inserted.
In scr_add_lines, in the case when a 0 width character must be inserted
in the first column of the current row, the calculation of the previous
character (i.e., the last character of the previous row) is bogus. In
particular, if the number of columns is 1, the 'tp' pointer points to
either the first or third byte (depending on the size of text_t) of the
rendition of the previous character, which is therefore corrupted by the
assignment '*tp = n'. This triggers a crash if the font id in the
rendition gets changed to the id of a font which is not loaded. Bug
reported by Kuang-che Wu.