Fix mouse reporting of motion notify events
On motion notify events, when no mouse button is pressed, urxvt
encodes either the number of the last button pressed or an invalid
value of -1, before the first button press, in mouse reporting
sequences. This commit fixes urxvt to report 3 as the button number,
same as xterm. Bug reported by Uli Schlachter.